Made parameters.py module and notebook for LCF.
